chore(search): add diagnostic log to inspect fuse.js options
The "Precise" search mode is exhibiting incorrect behavior, likely due to a misconfiguration of the Fuse.js `threshold` option when extended search is active.

This commit introduces a temporary `console.log` to output the options object being passed to the Fuse.js constructor. This will allow us to verify our hypothesis by observing the configuration in real-time as the "Precise" toggle is changed.